requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiobotocore==2.11.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyconfig==0.13.0
anyio==4.2.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
backoff==2.2.1
beautifulsoup4==4.11.1
binpickle==0.3.4
bitsandbytes==0.42.0
blessed==1.20.0
boto3==1.34.34
botocore==1.34.34
bs4==0.0.1
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chumpy==0.70
click==8.1.7
colorlog==6.8.2
commonmark==0.9.1
contourpy==1.0.6
croniter==1.4.1
csr==0.4.3
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.7.1
docker==6.1.3
docstring-parser==0.15
editor==1.6.6
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
ffmpeg-python==0.2.0
filelock==3.13.1
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gpt4all==2.0.2
grpcio==1.60.1
h11==0.14.0
hjson==3.1.0
huggingface-hub==0.23.0
hydra-core==1.3.2
idna==3.4
implicit==0.6.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
inquirer==3.2.3
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
jsonargparse==4.27.4
kiwisolver==1.4.4
lazy_loader==0.3
lenskit==0.14.2
librosa==0.10.1
lightning==2.1.4
lightning-api-access==0.0.5
lightning-cloud==0.5.64
lightning-fabric==2.1.4
lightning-utilities==0.10.1
llvmlite==0.39.1
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.6.2
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.5
natsort==8.4.0
networkx==2.8.8
ninja==1.11.1.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
openpyxl==3.1.2
ordered-set==4.1.0
packaging==24.2
pandas==1.5.2
patsy==0.5.3
Pillow==9.3.0
pip-search==0.0.12
platformdirs==4.2.0
pooch==1.8.0
protobuf==4.25.2
psutil==5.9.4
py-cpuinfo==9.0.0
py4j==0.10.9.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.8.0
pymeshlab==2022.2.post3
pynvml==11.5.0
pyparsing==3.0.9
pyrender==0.1.45
pyspark==3.3.1
python-dateutil==2.8.2
python-multipart==0.0.7
pytorch-lightning==2.1.4
pytz==2022.6
PyYAML==6.0.1
readchar==4.0.5
redis==5.0.1
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
runs==1.2.2
s3fs==2023.12.2
s3transfer==0.10.0
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
seedbank==0.1.2
shortuuid==1.0.11
six==1.16.0
smplx==0.1.28
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.7
starlette==0.35.0
statsmodels==0.13.5
surprise==0.1
sympy==1.12
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
timm==1.0.3
tokenizers==0.10.3
torchmetrics==1.3.0.post0
tqdm==4.64.1
traitlets==5.14.1
transformers==4.10.0
trimesh==4.3.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typeshed-client==2.4.0
typing==3.7.4.3
typing_extensions==4.8.0
urllib3==1.26.13
uvicorn==0.27.0.post1
wcwidth==0.2.13
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.16.0
xmod==1.8.1
yarl==1.9.4
zipp==3.17.0
